pyproject.toml
[project]
name = "edgetam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.3.1",
    "torchvision>=0.18.1",
    "numpy>=1.24.4",
    "tqdm>=4.66.1",
    "hydra-core>=1.3.2",
    "iopath>=0.1.10",
    "pillow>=9.4.0",
    "gradio==4.44.0",
    "gradio_client==1.3.0",
    "gradio_image_prompter==0.1.0",
    "opencv-python==4.10.0.84",
    "moviepy==1.0.3",
    "pydantic==2.10.6",
    "timm==1.0.15",
    "eva-decord==0.6.1",
    "socksio>=1.0.0",
]

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
